Fix missing class issue with H2 version 2.0.202
With older versions H2 returned a proprietary  instance from .
We used to support the conversion of that to an .

With the most recent versions  is no longer part of the H2 driver and we only register the converter when we encounter older versions of H2.

Closes #1114
See https://github.com/h2database/h2database/pull/1359
